+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules/Utxo.hs
@@ -19,8 +19,6 @@ module Cardano.Ledger.Alonzo.Rules.Utxo (
   AlonzoUTXO,
   AlonzoUtxoPredFailure (..),
   AlonzoUtxoEvent (..),
-  utxoPredFailMaToAlonzo,
-  utxoPredFailShelleyToAlonzo,
   validateCollateralContainsNonADA,
   validateExUnitsTooBigUTxO,
   validateOutputTooBigUTxO,
@@ -78,7 +76,6 @@ import Cardano.Ledger.Core
 import Cardano.Ledger.Credential (Credential (..))
 import Cardano.Ledger.Rules.ValidationMode (
   Inject (..),
-  InjectMaybe (..),
   Test,
   runTest,
   runTestOnSignal,
@@ -690,36 +687,6 @@ instance
 -- =====================================================
 -- Injecting from one PredicateFailure to another
 
-fromShelleyFailure :: ShelleyUtxoPredFailure era -> Maybe (AlonzoUtxoPredFailure era)
-fromShelleyFailure = \case
-  Shelley.BadInputsUTxO ins -> Just $ BadInputsUTxO ins
-  Shelley.ExpiredUTxO {} -> Nothing -- Replaced with `OutsideValidityIntervalUTxO` in Allegra
-  Shelley.MaxTxSizeUTxO a m -> Just $ MaxTxSizeUTxO a m
-  Shelley.InputSetEmptyUTxO -> Just InputSetEmptyUTxO
-  Shelley.FeeTooSmallUTxO mf af -> Just $ FeeTooSmallUTxO mf af
-  Shelley.ValueNotConservedUTxO vc vp -> Just $ ValueNotConservedUTxO vc vp
-  Shelley.WrongNetwork n as -> Just $ WrongNetwork n as
-  Shelley.WrongNetworkWithdrawal n as -> Just $ WrongNetworkWithdrawal n as
-  Shelley.OutputTooSmallUTxO {} -> Nothing -- Updated in Allegra
-  Shelley.UpdateFailure {} -> Nothing -- Removed
-  Shelley.OutputBootAddrAttrsTooBig outs -> Just $ OutputBootAddrAttrsTooBig outs
-
-fromAllegraFailure :: AllegraUtxoPredFailure era -> Maybe (AlonzoUtxoPredFailure era)
-fromAllegraFailure = \case
-  Allegra.BadInputsUTxO {} -> Nothing -- Inherited from Shelley
-  Allegra.OutsideValidityIntervalUTxO vi slotNo -> Just $ OutsideValidityIntervalUTxO vi slotNo
-  Allegra.MaxTxSizeUTxO {} -> Nothing -- Inherited from Shelley
-  Allegra.InputSetEmptyUTxO -> Nothing -- Inherited from Shelley
-  Allegra.FeeTooSmallUTxO {} -> Nothing -- Inherited from Shelley
-  Allegra.ValueNotConservedUTxO {} -> Nothing -- Inherited from Shelley
-  Allegra.WrongNetwork {} -> Nothing -- Inherited from Shelley
-  Allegra.WrongNetworkWithdrawal {} -> Nothing -- Inherited from Shelley
-  Allegra.OutputTooSmallUTxO {} -> Nothing -- Updated
-  Allegra.UpdateFailure {} -> Nothing -- Removed
-  Allegra.OutputBootAddrAttrsTooBig {} -> Nothing -- Inherited from Shelley
-  Allegra.TriesToForgeADA -> Just TriesToForgeADA
-  Allegra.OutputTooBigUTxO {} -> Nothing -- Updated error reporting
-
 instance Inject (AlonzoUtxoPredFailure era) (AlonzoUtxoPredFailure era) where
   inject = id
 
@@ -733,54 +700,35 @@ instance
   Inject (PPUPPredFailure era) (PredicateFailure (EraRule "UTXOS" era)) =>
   Inject (AllegraUtxoPredFailure era) (AlonzoUtxoPredFailure era)
   where
-  inject = utxoPredFailMaToAlonzo
-
-utxoPredFailMaToAlonzo ::
-  Inject (PPUPPredFailure era) (PredicateFailure (EraRule "UTXOS" era)) =>
-  AllegraUtxoPredFailure era ->
-  AlonzoUtxoPredFailure era
-utxoPredFailMaToAlonzo (Allegra.BadInputsUTxO x) = BadInputsUTxO x
-utxoPredFailMaToAlonzo (Allegra.OutsideValidityIntervalUTxO vi slotNo) =
-  OutsideValidityIntervalUTxO vi slotNo
-utxoPredFailMaToAlonzo (Allegra.MaxTxSizeUTxO x y) = MaxTxSizeUTxO x y
-utxoPredFailMaToAlonzo Allegra.InputSetEmptyUTxO = InputSetEmptyUTxO
-utxoPredFailMaToAlonzo (Allegra.FeeTooSmallUTxO c1 c2) = FeeTooSmallUTxO c1 c2
-utxoPredFailMaToAlonzo (Allegra.ValueNotConservedUTxO vc vp) = ValueNotConservedUTxO vc vp
-utxoPredFailMaToAlonzo (Allegra.WrongNetwork x y) = WrongNetwork x y
-utxoPredFailMaToAlonzo (Allegra.WrongNetworkWithdrawal x y) = WrongNetworkWithdrawal x y
-utxoPredFailMaToAlonzo (Allegra.OutputTooSmallUTxO x) = OutputTooSmallUTxO x
-utxoPredFailMaToAlonzo (Allegra.UpdateFailure x) = UtxosFailure (inject x)
-utxoPredFailMaToAlonzo (Allegra.OutputBootAddrAttrsTooBig xs) =
-  OutputTooBigUTxO (map (0,0,) xs)
-utxoPredFailMaToAlonzo Allegra.TriesToForgeADA = TriesToForgeADA
-utxoPredFailMaToAlonzo (Allegra.OutputTooBigUTxO xs) = OutputTooBigUTxO (map (\x -> (0, 0, x)) xs)
+  inject = \case
+    Allegra.BadInputsUTxO x -> BadInputsUTxO x
+    Allegra.OutsideValidityIntervalUTxO vi slotNo -> OutsideValidityIntervalUTxO vi slotNo
+    Allegra.MaxTxSizeUTxO x y -> MaxTxSizeUTxO x y
+    Allegra.InputSetEmptyUTxO -> InputSetEmptyUTxO
+    Allegra.FeeTooSmallUTxO c1 c2 -> FeeTooSmallUTxO c1 c2
+    Allegra.ValueNotConservedUTxO vc vp -> ValueNotConservedUTxO vc vp
+    Allegra.WrongNetwork x y -> WrongNetwork x y
+    Allegra.WrongNetworkWithdrawal x y -> WrongNetworkWithdrawal x y
+    Allegra.OutputTooSmallUTxO x -> OutputTooSmallUTxO x
+    Allegra.UpdateFailure x -> UtxosFailure (inject x)
+    Allegra.OutputBootAddrAttrsTooBig xs -> OutputTooBigUTxO (map (0,0,) xs)
+    Allegra.TriesToForgeADA -> TriesToForgeADA
+    Allegra.OutputTooBigUTxO xs -> OutputTooBigUTxO (map (\x -> (0, 0, x)) xs)
 
 instance
   Inject (PPUPPredFailure era) (PredicateFailure (EraRule "UTXOS" era)) =>
   Inject (ShelleyUtxoPredFailure era) (AlonzoUtxoPredFailure era)
   where
-  inject = utxoPredFailShelleyToAlonzo
-
-utxoPredFailShelleyToAlonzo ::
-  Inject (PPUPPredFailure era) (PredicateFailure (EraRule "UTXOS" era)) =>
-  ShelleyUtxoPredFailure era ->
-  AlonzoUtxoPredFailure era
-utxoPredFailShelleyToAlonzo (Shelley.BadInputsUTxO ins) = BadInputsUTxO ins
-utxoPredFailShelleyToAlonzo (Shelley.ExpiredUTxO ttl current) =
-  OutsideValidityIntervalUTxO (ValidityInterval SNothing (SJust ttl)) current
-utxoPredFailShelleyToAlonzo (Shelley.MaxTxSizeUTxO a m) = MaxTxSizeUTxO a m
-utxoPredFailShelleyToAlonzo Shelley.InputSetEmptyUTxO = InputSetEmptyUTxO
-utxoPredFailShelleyToAlonzo (Shelley.FeeTooSmallUTxO mf af) = FeeTooSmallUTxO mf af
-utxoPredFailShelleyToAlonzo (Shelley.ValueNotConservedUTxO vc vp) = ValueNotConservedUTxO vc vp
-utxoPredFailShelleyToAlonzo (Shelley.WrongNetwork n as) = WrongNetwork n as
-utxoPredFailShelleyToAlonzo (Shelley.WrongNetworkWithdrawal n as) = WrongNetworkWithdrawal n as
-utxoPredFailShelleyToAlonzo (Shelley.OutputTooSmallUTxO x) = OutputTooSmallUTxO x
-utxoPredFailShelleyToAlonzo (Shelley.UpdateFailure x) = UtxosFailure (inject x)
-utxoPredFailShelleyToAlonzo (Shelley.OutputBootAddrAttrsTooBig outs) =
-  OutputTooBigUTxO (map (0,0,) outs)
-
-instance InjectMaybe (ShelleyUtxoPredFailure era) (AlonzoUtxoPredFailure era) where
-  injectMaybe = fromShelleyFailure
-
-instance InjectMaybe (AllegraUtxoPredFailure era) (AlonzoUtxoPredFailure era) where
-  injectMaybe = fromAllegraFailure
+  inject = \case
+    Shelley.BadInputsUTxO ins -> BadInputsUTxO ins
+    Shelley.ExpiredUTxO ttl current ->
+      OutsideValidityIntervalUTxO (ValidityInterval SNothing (SJust ttl)) current
+    Shelley.MaxTxSizeUTxO a m -> MaxTxSizeUTxO a m
+    Shelley.InputSetEmptyUTxO -> InputSetEmptyUTxO
+    Shelley.FeeTooSmallUTxO mf af -> FeeTooSmallUTxO mf af
+    Shelley.ValueNotConservedUTxO vc vp -> ValueNotConservedUTxO vc vp
+    Shelley.WrongNetwork n as -> WrongNetwork n as
+    Shelley.WrongNetworkWithdrawal n as -> WrongNetworkWithdrawal n as
+    Shelley.OutputTooSmallUTxO x -> OutputTooSmallUTxO x
+    Shelley.UpdateFailure x -> UtxosFailure (inject x)
+    Shelley.OutputBootAddrAttrsTooBig outs -> OutputTooBigUTxO (map (0,0,) outs)