+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
@@ -9,12 +10,11 @@ module Test.Cardano.Ledger.Conway.Imp (
   spec,
 ) where
 
+import Cardano.Ledger.BaseTypes (Inject)
 import Cardano.Ledger.Conway.Governance (ConwayGovState, GovState)
-import Cardano.Ledger.Conway.Rules (
-  ConwayGOV,
-  ConwayLEDGER,
- )
+import Cardano.Ledger.Conway.Rules (ConwayGovPredFailure (..))
 import Cardano.Ledger.Core (EraRule)
+import Control.State.Transition.Extended (PredicateFailure)
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Conway.Imp.EnactSpec as Enact
 import qualified Test.Cardano.Ledger.Conway.Imp.EpochSpec as Epoch
@@ -25,8 +25,7 @@ spec ::
   forall era.
   ( ConwayEraImp era
   , GovState era ~ ConwayGovState era
-  , EraRule "GOV" era ~ ConwayGOV era
-  , EraRule "LEDGER" era ~ ConwayLEDGER era
+  , Inject (ConwayGovPredFailure era) (PredicateFailure (EraRule "LEDGER" era))
   ) =>
   Spec
 spec = do